availability_monitor/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.16
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
attrs==25.3.0
    # via aiohttp
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via
    #   requests
    #   yarl
multidict==6.4.3
    # via
    #   aiohttp
    #   yarl
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.4.0
    # via requests
yarl==1.20.0
    # via aiohttp
